(docs) Update readme with warning note and bigger example of defaults
Declaring both `Logrotate::Conf['/etc/logrotate.conf]` and
`Class['logrotate']` will result in errors from Puppet as described
in #116 which is the result of not using the class correctly. Update
readme to add a not warning users not to do this and update the
example for `Class['logrotate']` to indicate that feature parity
with `Logrotate::Conf['/etc/logrotate.conf']` already exists.

Fixes #116