Better testing if su is required.
I have made the su boolean no longer optional, defaulting to false.

I have made su_user and su_group no longer optional, defaulting to root.
Previously this was set top optional/undef and lateron in the code it
was changed to a fixed string when if was undef anyway.

I have added the su boolean to the logrotate::config to be consistent.